% Auto-generated: do not edit by hand
\name{discordCrate}

\alias{discordCrate}

\title{DiscordCrate component}

\description{

}

\usage{
discordCrate(id=NULL, allChannelNotifications=NULL, avatar=NULL,
channel=NULL, color=NULL, defer=NULL,
embedNotificationTimeout=NULL, glyph=NULL, indicator=NULL,
location=NULL, notifications=NULL, server=NULL, shard=NULL,
timeout=NULL, username=NULL)
}

\arguments{
\item{id}{Character. }

\item{allChannelNotifications}{Logical. }

\item{avatar}{Character. }

\item{channel}{Character. }

\item{color}{Character. }

\item{defer}{Logical. }

\item{embedNotificationTimeout}{Numeric. }

\item{glyph}{Unnamed list. }

\item{indicator}{Logical. }

\item{location}{Unnamed list. }

\item{notifications}{Logical. }

\item{server}{Character. }

\item{shard}{Character. }

\item{timeout}{Numeric. }

\item{username}{Character. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

