% Auto-generated: do not edit by hand
\name{discordWidget}

\alias{discordWidget}

\title{DiscordWidget component}

\description{

}

\usage{
discordWidget(id=NULL, avatar=NULL, channel=NULL, className=NULL,
server=NULL, style=NULL, username=NULL)
}

\arguments{
\item{id}{Character. }

\item{avatar}{Character. }

\item{channel}{Character. }

\item{className}{Character. }

\item{server}{Character. }

\item{style}{Named list. }

\item{username}{Character. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

