% Auto-generated: do not edit by hand
\name{discordWidget}

\alias{discordWidget}

\title{DiscordWidget component}

\description{

}

\usage{
discordWidget(id=NULL, avatar=NULL, channel=NULL, className=NULL,
height=NULL, server=NULL, username=NULL, width=NULL)
}

\arguments{
\item{id}{Character. }

\item{avatar}{Character. }

\item{channel}{Character. }

\item{className}{Character. }

\item{height}{Character. }

\item{server}{Character. }

\item{username}{Character. }

\item{width}{Character. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

